Move HLS support to scorer-ui-kit/hls sub-entry
Addresses reviewer feedback that the previous attempt at marking
hls.js as external broke consumer builds: a literal import("hls.js")
in our published bundle forced every consumer's bundler (Vite 8 /
Rolldown, webpack, etc.) to resolve hls.js at build time, even for
code paths that were never reached. Tree-shaking happens after module
resolution, so the runtime try/catch never got a chance to handle a
missing peer dep.

Split the HLS-aware variant into a separate entry:

  scorer-ui-kit         -> main entry, no hls.js footprint
  scorer-ui-kit/hls     -> LineUIVideoHLS, statically imports hls.js

Refactor:
  - Extract shared LineUI overlay logic into LineUIVideoBase
    (private, internal). Both wrappers render this base.
  - LineUIVideo now sets video.src = src directly. Same behavior for
    .mp4/.webm; .m3u8 in non-Safari browsers no longer auto-loads
    hls.js (consumers should switch to LineUIVideoHLS for that).
  - LineUIVideoHLS lives at src/LineUIHls/, statically imports Hls
    from 'hls.js', runs the HLS attach + autoplay + error-recovery
    flow on a videoRef passed to LineUIVideoBase.
  - vite.config.ts: multi-entry build (index + hls). Rolldown extracts
    LineUIVideoBase into a shared chunk so consumers using both
    entries don't double-ship the overlay code.
  - package.json: new "./hls" export. hls.js stays in
    peerDependenciesMeta.optional so non-HLS consumers see no install
    warning; consumers using /hls without hls.js installed get a clean
    runtime error pointing at the missing peer dep.
  - Example LineHLSPage swapped to import LineUIVideoHLS from
    scorer-ui-kit/hls. All other pages unchanged.

Verified matrix:
  - Consumer that never imports /hls, no hls.js installed: builds
    cleanly, bundle has zero hls.js code, all main-entry features work.
  - Consumer importing /hls, hls.js installed: builds, browser plays
    HLS via blob MediaSource (1280x720 @ 19s into BipBop manifest).
  - Consumer importing /hls, hls.js missing: build still succeeds
    (Rolldown emits a runtime stub); error fires only when navigating
    to the HLS page.
