+++ packages/example/src/pages/LineHLSPage.tsx
@@ -6,7 +6,6 @@ import {
   Sidebar,
   SidebarBox,
   Layout,
-  LineUIVideo,
   Content,
   Logo,
   ButtonWithIcon,
@@ -14,6 +13,7 @@ import {
   Input,
   Label,
 } from 'scorer-ui-kit';
+import { LineUIVideoHLS } from 'scorer-ui-kit/hls';
 import styled from 'styled-components';
 import { LineUIOptions, LineUIVideoOptions } from 'scorer-ui-kit/dist/LineUI';
 import ExamplesFilename from '../components/ExamplesFilename';
@@ -192,7 +192,7 @@ const LineHLSPage: React.FC = () => {
 
       <Content $padBottom={false}>
         <LineSetContext.Provider value={{ state, dispatch }}>
-          <LineUIVideo
+          <LineUIVideoHLS
             key={src}
             options={options}
             onLineClick={selectLine}
