+++ packages/ui-lib/src/LineUIHls/index.ts
@@ -1,0 +1,4 @@
+import LineUIVideoHLS from './LineUIVideoHLS';
+
+export { LineUIVideoHLS };
+export type { LineUIProps } from '../LineUI/LineUIVideoBase';
